add jobs module
