+++ src/print/unix/mod.rs
@@ -1,2 +1,16 @@
+use std::{borrow, ffi};
+
 pub mod cups;
+pub mod jobs;
 pub mod native;
+
+/// Performs lossy conversion from a [`ffi::CStr`] into [`String`].
+/// The result is either a borrowed value or an owned value.
+unsafe fn cstr_to_str(ptr: *const ffi::c_char) -> borrow::Cow<'static, str> {
+	unsafe { ffi::CStr::from_ptr(ptr).to_string_lossy() }
+}
+/// Constructs an owned UTF-8 string from a valid pointer to a valid C-string.
+/// Invalid characters are replaced with the replacement character.
+unsafe fn cstr_to_string(ptr: *const ffi::c_char) -> String {
+	unsafe { cstr_to_str(ptr).into_owned() }
+}
